refactor: replace int type for double type in prizes
- use double type to support big values in return rate calculation

+++ src/main/kotlin/lotto/model/LottoService.kt
@@ -59,8 +59,8 @@ class LottoService() {
 
     fun calculateReturnRate(rankList: Map<Rank, Int>, purchaseAmount: Int): Double {
         val totalWinnings = rankList.entries.sumOf { (rank , count) ->
-            rank.prize.toLong() * count
+            rank.prize * count
         }
-        return totalWinnings.toDouble() / purchaseAmount * 100
+        return totalWinnings / purchaseAmount * 100
     }
 }
